+++ UIReact/UICore/_executor/_Markup/Text/Generator.ts
@@ -2,29 +2,38 @@ import * as ReactDOMServer from 'react-dom/server';
 import * as React from 'react';
 import { Control } from 'UICore/Base';
 import { IControlOptions } from 'UICommon/Base';
-import { TemplateOrigin, IControlConfig, TemplateFunction } from './interfaces';
-import {IGeneratorAttrs, TemplateResult} from '../Vdom/interfaces';
-import {Logger} from 'UICommon/Utils';
+import { IControlConfig } from './interfaces';
 import {
    CommonUtils as Common,
-   RequireHelper,
-   ConfigResolver,
-   Scope,
-   plainMerge,
-   Helper,
-   IGeneratorNameObject,
-   ITplFunction, IAttributes, VoidTags as voidElements, TAttributes
+   IAttributes,
+   VoidTags as voidElements,
+   TAttributes,
+   IGenerator
 } from 'UICommon/Executor';
-import {Attr} from 'UICommon/Executor';
-import {IWasabyEvent} from 'UICommon/_events/IEvents';
+import { Attr } from 'UICommon/Executor';
+import { IWasabyEvent } from 'UICommon/_events/IEvents';
+import { Generator } from '../Generator';
 
 
 
 /**
  * @author Тэн В.А.
  */
-export class GeneratorText {
-   generatorBase: Generator;
+export class GeneratorText extends Generator implements IGenerator {
+   protected calculateOptions(
+       resolvedOptionsExtended: IControlOptions,
+       config: IControlConfig,
+       events: Record<string, IWasabyEvent[]>,
+       name: string): IControlOptions {
+      return {
+         ...resolvedOptionsExtended,
+         ...{events}
+      }
+   }
+
+   createText(text) {
+      return text;
+   };
 
    createWsControl(
        origin: string | typeof Control,
@@ -32,28 +41,7 @@ export class GeneratorText {
        _: unknown,
        __: unknown,
        deps: any
-   ): React.ComponentElement<
-       IControlOptions,
-       Control<IControlOptions, object>
-       > {
-      const controlClassExtended: TemplateOrigin = resolveTpl(origin, {}, deps);
-      const controlClass = Common.fixDefaultExport(controlClassExtended) as typeof Control;
-
-      // todo временное решение только для поддержки юнит-тестов
-      // https://online.sbis.ru/opendoc.html?guid=a886b7c1-fda3-4594-b00d-b48f1185aaf8
-      if (Common.isCompound(controlClass)) {
-         const generatorCompatibleStr = 'View/_executorCompatible/_Markup/Compatible/GeneratorCompatible';
-         const GeneratorCompatible = requirejs(generatorCompatibleStr).GeneratorCompatible;
-         const generatorCompatible = new GeneratorCompatible();
-         const markup = generatorCompatible.createWsControl.apply(generatorCompatible, arguments);
-         const res = React.createElement('remove', {
-            dangerouslySetInnerHTML: {
-               __html: markup
-            }
-         });
-         //@ts-ignore
-         return res;
-      }
+   ): string {
       if (origin.name ===  "StartApplicationScript") {
          const start = `<div><script>document.addEventListener('DOMContentLoaded', function() { let steps = [ { deps: ['Env/Env', 'Application/Initializer', 'Application/Env', 'SbisEnvUI/Wasaby', 'UI/Base', 'UI/State', 'Application/State', 'Core/polyfill'], callback: function (Env, AppInit, AppEnv, EnvUIWasaby, UIBase, UIState, AppState) { window.startContextData = {AppData: new UIState.AppData({})}; Object.assign(Env.constants, window.wsConfig); require( ["UIDemo/Index"], function () { var sr = new AppState.StateReceiver(UIState.Serializer); AppInit.default(window.wsConfig, void 0, sr); UIBase.BootstrapStart({}, document.getElementById('wasaby-content')); }); if (Env.constants.isProduction) { console.log( '%c\tЭта функция браузера предназначена для разработчиков. tЕсли кто-то сказал вам скопировать и вставить что-то здесь, это мошенники.\t\tВыполнив эти действия, вы предоставите им доступ к своему аккаунту.\t', 'background: red; color: white; font-size: 22px; font-weight: bolder; text-shadow: 1px 1px 2px black;' ); } } } ]; if (false) { steps.unshift({ deps: ['Core/polyfill'], callback: function () {} }) } function startApplication(steps) { let step = steps.shift(); require(step.deps, function(){ step.callback.apply(this, arguments); if (steps.length) { startApplication(steps); } }) } startApplication(steps); });</script></div>`
          const props = {
@@ -63,95 +51,11 @@ export class GeneratorText {
          };
          return ReactDOMServer.renderToString(React.createElement('div', props));
       }
-      return ReactDOMServer.renderToString(React.createElement(controlClass, scope));
+      return ReactDOMServer.renderToString(this.createReactControl(origin, scope, _, __, deps));
    }
 
-   createControlNew(
-       type: 'wsControl' | 'template',
-       origin: TemplateOrigin,
-       attributes: Attr.IAttributes,
-       events: Record<string, IWasabyEvent[]>,
-       options: IControlOptions,
-       config: IControlConfig
-   ): any {
-     const templateAttributes: IGeneratorAttrs = {
-        attributes: config.compositeAttributes === null
-            ? attributes
-            : Helper.processMergeAttributes(config.compositeAttributes, attributes),
-        /*
-        FIXME: https://online.sbis.ru/opendoc.html?guid=f354360c-5899-4f74-bf54-a06e526621eb
-        судя по нашей кодогенерации, createTemplate - это приватный метод, потому что она его не выдаёт.
-        Если это действительно так, то можно передавать родителя явным образом, а не через такие костыли.
-        Но т.к. раньше parent прокидывался именно так, то мне страшно это менять.
-        */
-        internal: {
-           parent: config.viewController
-        }
-     };
 
-     // вместо опций может прилететь функция, выполнение которой отдаст опции, calculateScope вычисляет такие опции
-     const resolvedOptions = Scope.calculateScope(options, plainMerge);
-     // если контрол создается внутри контентной опции, нужно пробросить в опции еще те, что доступны в контентной
-     // опции.
-     const resolvedOptionsExtended = ConfigResolver.addContentOptionScope(resolvedOptions, config);
-
-     /* FIXME: для wasabyOverReact, событий нет в options,
-     *   получить их можно только из data (объявляется в шаблонной функции)
-     *   такое поведение очень похоже на ошибку, т.к. все события должны быть в опциях
-     *   надо разобраться почему события в опции не попадают и убрать мерж опций из data
-     *   https://online.sbis.ru/opendoc.html?guid=c0aa021f-bd67-4fe9-8cfa-feee417fb3a3
-     */
-     const newOptions = {
-        ...resolvedOptionsExtended,
-        ...{events}
-     };
-
-     return this.resolver(origin, newOptions, templateAttributes, undefined,
-         config.depsLocal, config.includedTemplates);
-   }
-
-   resolver(
-       tplOrigin: TemplateOrigin,
-       preparedScope: IControlOptions,
-       decorAttribs: IGeneratorAttrs,
-       _: string,
-       deps?: Common.Deps<typeof Control, TemplateFunction>,
-       includedTemplates?: Common.IncludedTemplates<TemplateFunction>
-   ): React.ReactElement | React.ReactElement[] | string {
-      const parent = decorAttribs.internal.parent;
-
-      const tplExtended: TemplateOrigin = resolveTpl(tplOrigin, includedTemplates, deps);
-      const tpl = Common.fixDefaultExport(tplExtended);
-
-      // typeof Control
-      if (Common.isControlClass<typeof Control>(tpl)) {
-         return this.createWsControl(tpl, preparedScope, decorAttribs, undefined, deps);
-      }
-      // TemplateFunction - wml шаблон
-      if (Common.isTemplateClass<TemplateFunction>(tpl)) {
-         return this.createTemplate(tpl, preparedScope, decorAttribs, undefined, deps);
-      }
-      // inline template, xhtml, tmpl шаблон (closured), content option
-      if (typeof tpl === 'function') {
-         return resolveTemplateFunction(parent, tpl, preparedScope, decorAttribs);
-      }
-      // content option - в определенном способе использования контентная опция может представлять собой объект
-      // со свойством func, в котором и лежит функция контентной опции. Демка UITest/MarkupSpecification/resolver/Top
-      if (typeof tpl.func === 'function') {
-         return resolveTemplateFunction(parent, tpl.func, preparedScope, decorAttribs);
-      }
-
-      // Common.ITemplateArray - массив шаблонов, может например прилететь,
-      // если в контентной опции несколько корневых нод
-      if (Common.isTemplateArray<TemplateFunction>(tpl)) {
-         return resolveTemplateArray(parent, tpl, preparedScope, decorAttribs);
-      }
-
-      // не смогли зарезолвить - нужно вывести ошибку
-      logResolverError(tplOrigin, parent);
-      return '' + tplOrigin;
-   }
-   joinElements(elements: React.ReactNode): React.ReactNode {
+   joinElements(elements: string[]): string {
       if (Array.isArray(elements)) {
          let res = '';
          const self = this;
@@ -167,9 +71,8 @@ export class GeneratorText {
          throw new Error('joinElements: elements is not array');
       }
    }
-   createText(text) {
-      return text;
-   };
+
+
    createTag(tag, attrs, children, attrToDecorate?, defCollection?): string {
       if (!attrToDecorate) {
          attrToDecorate = {};
@@ -199,8 +102,15 @@ export class GeneratorText {
          return '<' + tag + mergedAttrsStr + ' />';
       }
       return '<' + tag + mergedAttrsStr + '>' + this.joinElements(children) + '</' + tag + '>';
-
    }
+
+   createDirective(text: string): string {
+      return '<' + text + '>';
+   };
+
+   escape<T>(value: T): T {
+      return Common.escape(value);
+   };
 }
 
 function decorateAttrs(attr1: TAttributes, attr2: TAttributes): string {
@@ -222,113 +132,4 @@ function decorateAttrs(attr1: TAttributes, attr2: TAttributes): string {
       return str;
    };
    return attrToStr(Attr.joinAttrs(attr1, attr2));
-}
-
-function resolveTemplateArray(
-    parent: Control<IControlOptions>,
-    templateArray: Common.ITemplateArray<TemplateFunction | ITplFunction<TemplateFunction>>,
-    resolvedScope: IControlOptions,
-    decorAttribs: IGeneratorAttrs): TemplateResult[] {
-   let result = [];
-   templateArray.forEach((template: TemplateFunction | ITplFunction<TemplateFunction>) => {
-      const resolvedTemplate = resolveTemplate(template, parent, resolvedScope, decorAttribs);
-      if (Array.isArray(resolvedTemplate)) {
-         result = result.concat(resolvedTemplate);
-      } else if (resolvedTemplate) {
-         result.push(resolvedTemplate);
-      }
-   });
-   return result;
-}
-
-function resolveTemplate(template: TemplateFunction | ITplFunction<TemplateFunction>,
-                         parent: Control<IControlOptions>,
-                         resolvedScope: IControlOptions,
-                         decorAttribs: IGeneratorAttrs): TemplateResult {
-   let resolvedTemplate;
-   if (typeof template === 'function') {
-      resolvedTemplate = resolveTemplateFunction(parent, template, resolvedScope, decorAttribs);
-   } else if (typeof template.func === 'function') {
-      resolvedTemplate = resolveTemplateFunction(parent, template.func, resolvedScope, decorAttribs);
-   } else {
-      resolvedTemplate = template;
-   }
-   if (Array.isArray(resolvedTemplate)) {
-      if (resolvedTemplate.length === 1) {
-         return resolvedTemplate[0];
-      }
-      if (resolvedTemplate.length === 0) {
-         // return null so that resolveTemplateArray does not add
-         // this to the result array, since it is empty
-         return null;
-      }
-   }
-   return resolvedTemplate;
-}
-
-function anonymousFnError(fn: TemplateFunction, parent: Control<IControlOptions>): void {
-   Logger.error(`Ошибка построения разметки. Была передана функция, которая не является шаблонной.
-               Функция: ${fn.toString()}`, parent);
-}
-
-function logResolverError(tpl: TemplateOrigin, parent: Control<IControlOptions>): void {
-   if (typeof tpl !== 'string') {
-      let errorText = 'Ошибка в шаблоне! ';
-      if (Common.isLibraryModule(tpl)) {
-         errorText += `Контрол не найден в библиотеке.
-                Библиотека: ${(tpl as IGeneratorNameObject).library}.
-                Контрол: ${(tpl as IGeneratorNameObject).module}`;
-      } else {
-         errorText += `Неверное значение в ws:partial. Шаблон: ${tpl} имеет тип ${typeof tpl}`;
-      }
-      Logger.error(errorText, parent);
-   }
-   if (typeof tpl === 'string' && tpl.split('!')[0] === 'wml'){
-      // если у нас тут осталась строка то проверим не путь ли это до шаблона
-      // если это так, значит мы не смогли построить контрол, т.к. указан не существующий шаблон
-      Logger.error('Ошибка при построение контрола. Проверьте существует ли шаблон ' + tpl, parent);
-   }
-}
-
-function resolveTemplateFunction(parent: Control<IControlOptions>,
-                                 template: TemplateFunction,
-                                 resolvedScope: IControlOptions,
-                                 decorAttribs: IGeneratorAttrs): TemplateResult {
-   if (Common.isAnonymousFn(template)) {
-      anonymousFnError(template, parent);
-      return null;
-   }
-   return template.call(parent, resolvedScope, decorAttribs, undefined, true, undefined, undefined) as TemplateResult;
-}
-
-function getLibraryTpl(tpl: IGeneratorNameObject,
-                       deps: Common.Deps<typeof Control, TemplateFunction>
-): typeof Control | Common.ITemplateArray<TemplateFunction> {
-   let controlClass;
-   if (deps && deps[tpl.library]) {
-      controlClass = Common.extractLibraryModule(deps[tpl.library], tpl.module);
-   } else if (RequireHelper.defined(tpl.library)) {
-      controlClass = Common.extractLibraryModule(RequireHelper.extendedRequire(tpl.library, tpl.module), tpl.module);
-   }
-   return controlClass;
-}
-function resolveTpl(tpl: TemplateOrigin,
-                    includedTemplates: Common.IncludedTemplates<TemplateFunction>,
-                    deps: Common.Deps<typeof Control, TemplateFunction>
-): typeof Control | TemplateFunction | Common.IDefaultExport<typeof Control> |
-    Function | Common.ITemplateArray<TemplateFunction> {
-   if (typeof tpl === 'string') {
-      if (Common.isLibraryModuleString(tpl)) {
-         // if this is a module string, it probably is from a dynamic partial template
-         // (ws:partial template="{{someString}}"). Split library name and module name
-         // here and process it in the next `if tpl.library && tpl.module`
-         const tplObject = Common.splitModule(tpl);
-         return getLibraryTpl(tplObject, deps);
-      }
-      return Common.depsTemplateResolver(tpl, includedTemplates, deps);
-   }
-   if (Common.isLibraryModule(tpl)) {
-      return getLibraryTpl(tpl, deps);
-   }
-   return tpl;
 }
